Added Rule Types
Added RuleType, FunctionRule, and RuleDuo. RuleType is the type of the boolean function. Similar to the RuleType enum from Webster. FunctionRule contains the name of the variable and the RuleType. RuleDuo simply contains two rules. This is almost an exact port of Webster's Rule system, but now the value is contained inside the RuleType enum.